+++ src/lib/libast/sfio/sfresize.c
@@ -45,12 +45,12 @@ int sfresize(Sfio_t* f, Sfoff_t size)
 					f->extent = s;
 				}
 			}
-			memclear((char*)(f->data+size), (int)(f->extent-size));
+			memclear((char*)(f->data+size), f->extent-size);
 		}
 		else
 		{	if(SFSK(f, size, SEEK_SET, f->disc) != size)
 				return -1;
-			memclear((char*)(f->data+f->extent), (int)(size-f->extent));
+			memclear((char*)(f->data+f->extent), size-f->extent);
 		}
 	}
 	else
